chore: examples tested and fixes
